+++ enemy.h
@@ -8,7 +8,7 @@ class Enemy : public Character
 private:
 public:
     Enemy(int x_pos, int y_pos, QString img_file, int health, int power, int speed, int r);
-    virtual void updateFrame(int frame);
+    virtual void updateFrame(long long frame);
 
     ~Enemy();
