+++ explosion.h
@@ -9,7 +9,7 @@ private:
     int current;
 public:
     Explosion(float x_pos, float y_pos);
-    void updateFrame(int frame);
+    void updateFrame(long long frame);
     ~Explosion();
 };
